requirements.txt
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
albumentations>=1.3.1
numpy>=1.24.0
PyYAML>=6.0
wandb>=0.16.0
tqdm>=4.65.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
onnx>=1.15.0
onnxruntime>=1.16.0
editdistance>=0.6.3
